COZMO-15195: Feeding Cubes draining too quickly
Switch ints to floats to increase precision
All other fixes included in anims DEPS changes

+++ engine/behaviorSystem/activities/activities/activityFeeding.cpp
@@ -713,7 +713,7 @@ IBehaviorPtr ActivityFeeding::GetBestBehaviorFromMap() const
 
 // Implementation of IFeedingListener
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-void ActivityFeeding::StartedEating(Robot& robot, const int duration_s)
+void ActivityFeeding::StartedEating(Robot& robot, const float duration_s)
 {
   using CS = FeedingCubeController::ControllerState;
   _cubeControllerMap[_cubeIDToEat]->SetControllerState(robot, CS::DrainCube, duration_s);
